change directory and split definition and implement
